+++ CHANGELOG.md
@@ -209,6 +209,7 @@
 - In manufacturing order, Stock move for produced product now uses the price from the product instead of from the costsheet.
 - Manuf Order : Fix forbid to set a value equal to zero on wizard when we click on 'Updated planned quantities (Components and FP)' button.
 - Manuf Order & ProdProcess : Added canEdit false on company and stockLocations fields and used widget SuggestBox for company.
+- Invoice : fix fiscal position of account for partner account.
 
 ## [Unreleased 5.0.7]
 ## Features